Moving splits to eval/predict/splits.py
